requirements.txt
PyYAML
yfinance
rich
fastapi
jinja2
uvicorn